Add downloadTo to the Storage interface
