+++ google-cloud-storage/src/test/java/com/google/cloud/storage/BlobTest.java
@@ -612,102 +612,4 @@ public class BlobTest {
     assertNull(blob.getCustomTime());
     assertTrue(blob.isDirectory());
   }
-
-  private StorageRpc prepareForDownload() {
-    StorageRpc mockStorageRpc = createNiceMock(StorageRpc.class);
-    expect(storage.getOptions()).andReturn(mockOptions).anyTimes();
-    replay(storage);
-    expect(mockOptions.getStorageRpcV1()).andReturn(mockStorageRpc);
-    expect(mockOptions.getRetrySettings()).andReturn(RETRY_SETTINGS);
-    expect(mockOptions.getClock()).andReturn(API_CLOCK);
-    expect(mockOptions.getRetryAlgorithmManager()).andReturn(retryAlgorithmManager).anyTimes();
-    replay(mockOptions);
-    blob = new Blob(storage, new BlobInfo.BuilderImpl(BLOB_INFO));
-    return mockStorageRpc;
-  }
-
-  @Test
-  public void testDownloadTo() throws Exception {
-    final byte[] expected = {1, 2};
-    StorageRpc mockStorageRpc = prepareForDownload();
-    expect(
-            mockStorageRpc.read(
-                anyObject(StorageObject.class),
-                anyObject(Map.class),
-                eq(0l),
-                anyObject(OutputStream.class)))
-        .andAnswer(
-            new IAnswer<Long>() {
-              @Override
-              public Long answer() throws Throwable {
-                ((OutputStream) getCurrentArguments()[3]).write(expected);
-                return 2l;
-              }
-            });
-    replay(mockStorageRpc);
-    File file = File.createTempFile("blob", ".tmp");
-    blob.downloadTo(file.toPath());
-    byte actual[] = Files.readAllBytes(file.toPath());
-    assertArrayEquals(expected, actual);
-  }
-
-  @Test
-  public void testDownloadToWithRetries() throws Exception {
-    final byte[] expected = {1, 2};
-    StorageRpc mockStorageRpc = prepareForDownload();
-    expect(
-            mockStorageRpc.read(
-                anyObject(StorageObject.class),
-                anyObject(Map.class),
-                eq(0l),
-                anyObject(OutputStream.class)))
-        .andAnswer(
-            new IAnswer<Long>() {
-              @Override
-              public Long answer() throws Throwable {
-                ((OutputStream) getCurrentArguments()[3]).write(expected[0]);
-                throw new StorageException(504, "error");
-              }
-            });
-    expect(
-            mockStorageRpc.read(
-                anyObject(StorageObject.class),
-                anyObject(Map.class),
-                eq(1l),
-                anyObject(OutputStream.class)))
-        .andAnswer(
-            new IAnswer<Long>() {
-              @Override
-              public Long answer() throws Throwable {
-                ((OutputStream) getCurrentArguments()[3]).write(expected[1]);
-                return 1l;
-              }
-            });
-    replay(mockStorageRpc);
-    File file = File.createTempFile("blob", ".tmp");
-    blob.downloadTo(file.toPath());
-    byte actual[] = Files.readAllBytes(file.toPath());
-    assertArrayEquals(expected, actual);
-  }
-
-  @Test
-  public void testDownloadToWithException() throws Exception {
-    StorageRpc mockStorageRpc = prepareForDownload();
-    Exception exception = new IllegalStateException("test");
-    expect(
-            mockStorageRpc.read(
-                anyObject(StorageObject.class),
-                anyObject(Map.class),
-                eq(0l),
-                anyObject(OutputStream.class)))
-        .andThrow(exception);
-    replay(mockStorageRpc);
-    File file = File.createTempFile("blob", ".tmp");
-    try {
-      blob.downloadTo(file.toPath());
-      fail();
-    } catch (StorageException e) {
-      assertSame(exception, e.getCause());
-    }
-  }
 }
