+++ google-cloud-storage/src/test/java/com/google/cloud/storage/it/ITStorageTest.java
@@ -883,17 +883,14 @@ public class ITStorageTest {
     storage.createFrom(blobInfo, gzippedFile.toPath());
 
     Path rawInputGzippedFile = File.createTempFile("rawinputgzippedfile", ".txt").toPath();
-    Blob blob = storage.get(blobId);
 
-    blob.downloadTo(rawInputGzippedFile, Blob.BlobSourceOption.shouldReturnRawInputStream(true));
+    storage.downloadTo(blobId, rawInputGzippedFile, Blob.BlobSourceOption.shouldReturnRawInputStream(true));
 
     assertArrayEquals(
         Files.readAllBytes(gzippedFile.toPath()), Files.readAllBytes(rawInputGzippedFile));
 
     Path unzippedFile = File.createTempFile("unzippedfile", ".txt").toPath();
-    storage
-        .get(blobId)
-        .downloadTo(unzippedFile, Blob.BlobSourceOption.shouldReturnRawInputStream(false));
+    storage.downloadTo(blobId, unzippedFile, Blob.BlobSourceOption.shouldReturnRawInputStream(false));
 
     assertArrayEquals("hello world".getBytes(), Files.readAllBytes(unzippedFile));
   }
@@ -3686,7 +3683,7 @@ public class ITStorageTest {
     assertEquals(BLOB_BYTE_CONTENT.length, (long) blob.getSize());
 
     Path tempFileTo = Files.createTempFile("ITStorageTest_", ".tmp");
-    storage.get(blobId).downloadTo(tempFileTo);
+    storage.downloadTo(blobId, tempFileTo);
     byte[] readBytes = Files.readAllBytes(tempFileTo);
     assertArrayEquals(BLOB_BYTE_CONTENT, readBytes);
   }
